fix: move utilities.js to Product folder
